Updated slack.rb with send msg menu methods
