+++ lib/workspace.rb
@@ -47,9 +47,38 @@ class Workspace
     return "name: #{x.name} | topic: #{x.topic} | Slack id: #{x.slack_id}"
   end
   
-  def send_message(name, message)
-    x = select_channel(name)
-    
+  def send_msg_to_user(user, message)
+    x = select_user(user)
+    url = BASE_URL + "chat.postMessage"
+    params = {
+      token: KEY,
+      channel: x.slack_id,
+      text: message,
+    }
+
+    posted_message = HTTParty.post(url, query: params)
+    if posted_message["ok"] == false
+      raise SlackApiError, "API call failed with reason #{posted_message["error"]}"
+    end
+
+    return "You: #{posted_message["message"]["text"]}"
+  end
+
+  def send_msg_to_channel(channel, message)
+    x = select_channel(channel)
+    url = BASE_URL + "chat.postMessage"
+    params = {
+      token: KEY,
+      channel: x.slack_id,
+      text: message,
+    }
+
+    posted_message = HTTParty.post(url, query: params)
+    if posted_message["ok"] == false
+      raise SlackApiError, "API call failed with reason #{posted_message["error"]}"
+    end
+
+    return "You: #{posted_message["message"]["text"]}"
   end
 end
